rril: 64 bits build fixes
Correct pointer to int casting errors.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-9079
Reviewed-on: https://android.intel.com:443/300436

+++ CORE/ND/MODEMS/initializer.cpp
@@ -598,7 +598,7 @@ BOOL CInitializer::CreateChannels()
         if (!g_pRilChannel[i] || !g_pRilChannel[i]->Initialize())
         {
             RIL_LOG_CRITICAL("CInitializer::CreateChannels() : Channel[%d] (0x%X)"
-                    " Init failed\r\n", i, (UINT32)g_pRilChannel[i]);
+                    " Init failed\r\n", i, (intptr_t)g_pRilChannel[i]);
             goto Error;
         }
